Finished I think? Time to clean up

+++ includes/handle_admin.php
@@ -4,11 +4,15 @@
 	{
 		$id = $_POST["id"];
 
-		// Update that user's status to unlocked
-		$stmt = "UPDATE Users SET locked='true' WHERE id=$id;";
-		mysqli_query($db, $stmt);
+		// Assignment spec says you can only lock regular users
+		if ($id != 1)
+		{
+			// Update that user's status to unlocked
+			$stmt = "UPDATE Users SET locked='true' WHERE id=$id;";
+			mysqli_query($db, $stmt);
 
-		header("Location: /admin.php");
+			header("Location: /admin.php");
+		}
 	}
 	// If a post request was made to unlock the account
 	else if (isset($_POST["unlock"]))
@@ -27,10 +31,14 @@
 	{
 		$id = $_POST["id"];
 
-		// Delete the account
-		$stmt = "DELETE FROM Users WHERE id=$id;";
-		mysqli_query($db, $stmt);
-
-		header("Location: /admin.php");
+		// Do not allow admins to delete the admin account
+		if ($id != 1)
+		{
+			// Delete the account
+			$stmt = "DELETE FROM Users WHERE id=$id;";
+			mysqli_query($db, $stmt);
+	
+			header("Location: /admin.php");
+		}
 	}
 ?>
